Use u32 QUIC error_code
The error code is 62-bit in QUIC. In this API, limit to 32 bits:
 - it avoids complexity to check that a u64 value fits into 62 bits;
 - in JavaScript, numbers are float, so they can store only 52-bit
   precise integers;
 - 32 bits is more than enough for our use case.

+++ src/driver/mod.rs
@@ -18,7 +18,7 @@ pub trait ConnectionDriver {
 
     async fn send_datagram(&self, data: Bytes) -> Result<(), SendDatagramError>;
 
-    fn close(&self, error_code: u64, reason: &str);
+    fn close(&self, error_code: u32, reason: &str);
 
     fn max_datagram_size(&self) -> Option<usize>;
 }
